ml/pyproject.toml
[project]
name = "wholabass-ml"
version = "0.1.0"
description = "ML sidecar for wholabass"
requires-python = ">=3.11,<3.13"
dependencies = [
  "demucs>=4.0.1",
  "soundfile>=0.13",
  "torchcodec>=0.4",
  "yt-dlp>=2024.5.27",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "mypy>=1.10",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["server.py", "progress.py", "pipeline"]

[[tool.mypy.overrides]]
module = ["demucs.*", "soundfile", "torchaudio", "torchaudio.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
